Long/ULong Conversion Rework (Issue #156) (#199)
Changed all manual casting between long and ulong in Modix.Data over to have EF do it for us, via ModelBuilder.Entity().Property().HasConversion(). This resolves #156.

+++ Modix.Data/Models/Core/DesignatedChannelMappingCreationData.cs
@@ -30,15 +30,15 @@ namespace Modix.Data.Models.Core
         internal DesignatedChannelMappingEntity ToEntity()
             => new DesignatedChannelMappingEntity()
             {
-                GuildId = (long)GuildId,
-                ChannelId = (long)ChannelId,
+                GuildId = GuildId,
+                ChannelId = ChannelId,
                 Type = Type,
                 CreateAction = new ConfigurationActionEntity()
                 {
-                    GuildId = (long)GuildId,
+                    GuildId = GuildId,
                     Type = ConfigurationActionType.DesignatedChannelMappingCreated,
                     Created = DateTimeOffset.Now,
-                    CreatedById = (long)CreatedById
+                    CreatedById = CreatedById
                 }
             };
     }
