+++ Modix.Data/Models/Core/GuildUserBrief.cs
@@ -40,7 +40,7 @@ namespace Modix.Data.Models.Core
         internal static readonly Expression<Func<GuildUserEntity, GuildUserBrief>> FromEntityProjection
             = entity => new GuildUserBrief()
             {
-                Id = (ulong)entity.User.Id,
+                Id = entity.User.Id,
                 Username = entity.User.Username,
                 Discriminator = entity.User.Discriminator,
                 Nickname = entity.Nickname
