+++ Modix.Data/Models/Promotions/PromotionCampaignDetails.cs
@@ -59,7 +59,7 @@ namespace Modix.Data.Models.Promotions
             = entity => new PromotionCampaignDetails()
             {
                 Id = entity.Id,
-                GuildId = (ulong)entity.GuildId,
+                GuildId = entity.GuildId,
                 Subject = entity.Subject.Project(GuildUserBrief.FromEntityProjection),
                 TargetRole = entity.TargetRole.Project(GuildRoleBrief.FromEntityProjection),
                 Outcome = entity.Outcome,
